+++ iqmotion/__init__.py
@@ -1,9 +1,15 @@
-from iqmotion.iq_devices.custom_iq_module import CustomIqModule
+from iqmotion.iq_devices.base_iq_module import BaseIqModule
 
+# Soon to be deprecated: but the original way to access motor firmware API
+from iqmotion.iq_devices.custom_iq_module import CustomIqModule
 from iqmotion.iq_devices.speed_module import SpeedModule
 from iqmotion.iq_devices.servo_module import ServoModule
 from iqmotion.iq_devices.step_dir_module import StepDirModule
 
+# Make Motor Modules front facing visible
+from iqmotion.iq_devices.vertiq8108_module import Vertiq8108
+from iqmotion.iq_devices.vertiq2306_module import Vertiq2306
+from iqmotion.iq_devices.fortiq_module  import Fortiq
 
 from iqmotion.communication.communicator import Communicator
 from iqmotion.communication.serial_communicator import SerialCommunicator